feat: If an object is handed to addLayer, pass it through to Mapbox.
There has been some confusion due to the fact that Mapbox GL JS also has an
addLayer method in its API, but it is only useable in the AirMap SDK by calling
`map.mapbox.addLayer`. To simplify things, we will now call Mapbox's
`addLayer` if an object is passed to this method. If a string is passed,
everything will continue to behave as expected.